Create TabsMenu components
