--- src/ui/layout/menu/components/NumberOfMyTeamWorkOrders.jsx
@@ -1,11 +1,0 @@
-import React from 'react';
-import NumberBadge from './NumberBadge'; 
-import useMyTeamWorkOrdersStore from '../../../../state/useMyTeamWorkOrdersStore';
-
-const NumberOfMyTeamWorkOrders = () => {
-  const { myTeamWorkOrders } = useMyTeamWorkOrdersStore();
-
-  return <NumberBadge count={myTeamWorkOrders.length} />;
-};
-
-export default NumberOfMyTeamWorkOrders;
